MyApps-13: Move web cache implementation, api and dependency injection to submodules.
