+++ app/src/main/java/xyz/torquato/myapps/di/data/DataModule.kt
@@ -5,9 +5,7 @@ import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import xyz.torquato.myapps.api.math.IMathRepository
-import xyz.torquato.myapps.api.web.cache.IWebCacheRepository
 import xyz.torquato.myapps.data.math.MathRepository
-import xyz.torquato.myapps.data.web.cache.WebCacheRepository
 import javax.inject.Singleton
 
 @Module
@@ -20,9 +18,7 @@ abstract class RepositoryModule {
 
 
 
-    @Binds
-    @Singleton
-    abstract fun bindWebCacheImpl(impl: WebCacheRepository): IWebCacheRepository
+
