+++ app/src/main/java/xyz/torquato/myapps/domain/web/content/GetCacheBooksUseCase.kt
@@ -2,10 +2,9 @@ package xyz.torquato.myapps.domain.web.content
 
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
-import xyz.torquato.myapps.api.web.cache.CacheItem
-import xyz.torquato.myapps.api.web.cache.IWebCacheRepository
+import xyz.torquato.myaps.domain.api.web.cache.CacheItem
+import xyz.torquato.myaps.domain.api.web.cache.IWebCacheRepository
 import xyz.torquato.myaps.domain.api.web.model.BookItem
-import kotlin.collections.map
 
 class GetCacheBooksUseCase(
     val webCacheRepository: IWebCacheRepository
