+++ data/di/src/main/java/xyz/torquato/myapps/data/di/RepositoryImpl.kt
@@ -10,12 +10,14 @@ import xyz.torquato.myapps.data.impl.waves.SoundRepository
 import xyz.torquato.myapps.data.impl.web.LocalWebRepository
 import xyz.torquato.myapps.data.impl.web.QueryRepository
 import xyz.torquato.myapps.data.impl.web.WebRepository
+import xyz.torquato.myapps.data.impl.web.cache.WebCacheRepository
 import xyz.torquato.myaps.domain.api.encrypt.ICypherRepository
 import xyz.torquato.myaps.domain.api.pagging.IPagingRepository
 import xyz.torquato.myaps.domain.api.sound.ISoundRepository
 import xyz.torquato.myaps.domain.api.web.ILocalWebRepository
 import xyz.torquato.myaps.domain.api.web.IQueryRepository
 import xyz.torquato.myaps.domain.api.web.IWebRepository
+import xyz.torquato.myaps.domain.api.web.cache.IWebCacheRepository
 import javax.inject.Singleton
 
 @Module
@@ -44,5 +46,9 @@ abstract class RepositoryImpl {
     @Binds
     @Singleton
     abstract fun bindLocalWebImpl(impl: LocalWebRepository): ILocalWebRepository
+
+    @Binds
+    @Singleton
+    abstract fun bindWebCacheImpl(impl: WebCacheRepository): IWebCacheRepository
 
 }
