+++ data/impl/src/main/java/xyz/torquato/myapps/data/impl/web/cache/WebCacheDataSource.kt
@@ -1,8 +1,8 @@
-package xyz.torquato.myapps.data.web.cache
+package xyz.torquato.myapps.data.impl.web.cache
 
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import xyz.torquato.myapps.api.web.cache.CacheItem
+import xyz.torquato.myaps.domain.api.web.cache.CacheItem
 import javax.inject.Inject
 
 class WebCacheDataSource @Inject constructor() {
