+++ data/impl/src/main/java/xyz/torquato/myapps/data/impl/web/cache/WebCacheRepository.kt
@@ -1,13 +1,13 @@
-package xyz.torquato.myapps.data.web.cache
+package xyz.torquato.myapps.data.impl.web.cache
 
 import kotlinx.coroutines.flow.StateFlow
-import xyz.torquato.myapps.api.web.cache.CacheItem
-import xyz.torquato.myapps.api.web.cache.IWebCacheRepository
+import xyz.torquato.myaps.domain.api.web.cache.CacheItem
+import xyz.torquato.myaps.domain.api.web.cache.IWebCacheRepository
 import javax.inject.Inject
 
 class WebCacheRepository @Inject constructor(
     val dataSource: WebCacheDataSource,
-) : IWebCacheRepository{
+) : IWebCacheRepository {
 
     override val cache: StateFlow<List<CacheItem>> = dataSource.cache
